Add ability to report checker failures, better .pylintrc integration
Fixes #6. When a checker fails, and we can detect it, we can now report the
failure inline with other errors, which should help in a) noticing the failure
in the first place, and b) debugging it. This is configurable via the
`flycheck-pycheckers-report-errors-inline` variable.

The specific failure noticed here was that pylint doesn't know where its
`.pylintrc` file can be found within the context of the init-hook specified by
that same `.pylintrc` file (`find_pylintrc()` would fail), and so using this
config file would cause pylint to silently fail. Pylint is now passed
information about this config file via both the `--rcfile` option and the
`PYLINTRC` environment variable.