Filter non-applicable ignore codes per checker
Some checkers (bandit) crash when told to ignore a code they don't know about,
rather than ignoring it. This change filters the list of codes that are passed
in the `ignore` option to a hardcoded, per-checker list of acceptable
codes. It's a hack, but it lets bandit work with the default settings.