feat(mypy): Hint for files existing outside project directory
I am working in a project where all tests are placed in a sibling directory
next to the main project's source root.

cd ./some_project
tree
├── Makefile
├── README.md
├── some_project
│   ├── __init__.py
│   └── __main__.py
├── mypy.ini
├── pytest.ini
├── requirements-frozen.txt
├── requirements.txt
└── tests
    ├── conftest.py
    └── unit_tests
        ├── some_feature
        │   ├── test_foo.py
        │   └── test_bar.py
        └── test_baz.py

If you include a `mypy_path` in `mypy.ini` pointing to your current project,
when combined with `find_project_root`, a check can be made to use your
project's existing mypy type definitions while editing code in the `tests`
directory.

When editing code that matches a directory that shares its base with
a `mypy_path`, nothing is done since it is already in mypy's context.